+++ src/Skoruba.Duende.IdentityServer.Admin.UI.Client/src/pages/Client/Edit/Tabs/SecretsTab.tsx
@@ -11,10 +11,19 @@ import { useParams } from "react-router-dom";
 import { queryKeys } from "@/services/QueryKeys";
 import { Key } from "lucide-react";
 import { CardWrapper } from "@/components/CardWrapper/CardWrapper";
+import { Tip } from "@/components/Tip/Tip";
+import { ClientEditFormData } from "../../ClientSchema";
+import { useFormContext, useWatch } from "react-hook-form";
 
 const SecretsTab: React.FC = () => {
   const { t } = useTranslation();
   const { clientId } = useParams<{ clientId: string }>();
+  const { control } = useFormContext<ClientEditFormData>();
+
+  const requireClientSecret = useWatch({
+    control,
+    name: "requireClientSecret",
+  });
 
   return (
     <CardWrapper
@@ -35,6 +44,10 @@ const SecretsTab: React.FC = () => {
         type="switch"
         includeSeparator
       />
+
+      {!requireClientSecret && (
+        <Tip className="mb-4">{t("Client.Tabs.SecretsNotUsed")}</Tip>
+      )}
 
       <Secrets
         resourceId={Number(clientId)}